requirements.txt
# AI 开发团队依赖
# 安装: pip install -r requirements.txt

# 核心框架
crewai>=0.28.0
langchain-anthropic>=0.1.0

# GitHub集成
PyGithub>=2.1.0

# HTTP请求
requests>=2.31.0

# 环境变量
python-dotenv>=1.0.0

# Web服务器 (用于Webhook)
flask>=3.0.0

# 生产环境服务器 (Railway需要)
gunicorn>=21.0.0

# 类型检查 (可选)
mypy>=1.8.0
